# apps/docs/public/openapi-books.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Syncero Books API",
    "version": "1.0.0",
    "description": "API privada do **Syncero Books** — app para contadores.\n\n### Autenticação\n- **Rotas internas:** Bearer JWT emitido pelo Supabase (requer `Origin` de domínio Syncero)\n- **Rotas externas:** `x-api-key` (API Key do sistema) ou Bearer JWT\n\n### Restrição de Origem (Origin Guard)\nRotas internas verificam o header `Origin` em produção.\nApenas origens em `ALLOWED_ORIGINS` são aceitas — rotas externas não têm essa restrição.\n\n### Rotas externas (aceitam x-api-key)\n- `GET /api/transactions` — listar lançamentos de empresa externa\n- `POST /api/transactions` — criar lançamento para empresa externa\n- `POST /api/journal-entries` — classificar lançamento (criar partida dobrada)\n\n### Autorização interna\n- Empresas Flow: `accountant_companies.status = 'accepted'`\n- Empresas externas: `external_companies.accountant_id = userId`"
  },
  "servers": [
    {
      "url": "http://localhost:5175",
      "description": "Desenvolvimento local"
    },
    {
      "url": "https://syncero-books.vercel.app",
      "description": "Produção"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKey": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API Key gerada pelo sistema (GET/POST /api-keys). Aceita nas rotas externas como alternativa ao Bearer JWT."
      },
      "originGuard": {
        "type": "apiKey",
        "in": "header",
        "name": "Origin",
        "description": "Header de origem verificado nas rotas internas. Em produção, apenas origens Syncero são aceitas."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "JournalSide": {
        "type": "string",
        "enum": [
          "debit",
          "credit"
        ]
      },
...
      "JournalEntryLine": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "entry_id": {
            "type": "string",
            "format": "uuid"
          },
          "account_plan_id": {
            "type": "string",
            "format": "uuid"
          },
          "side": {
            "$ref": "#/components/schemas/JournalSide"
          },
          "amount": {
            "type": "number"
          },
          "memo": {
            "type": "string",
            "nullable": true
          },
          "account_plan": {
            "type": "object",
            "nullable": true,
            "properties": {
              "code": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          }
        }
      },
      "JournalEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "ext_company_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "accountant_id": {
            "type": "string",
            "format": "uuid"
          },
          "entry_date": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "enum": [
              "manual",
              "api",
              "syncero_import"
            ]
          },
          "flow_transaction_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "external_ref": {
            "type": "string",
            "nullable": true
          },
          "is_reversed": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JournalEntryLine"
            }
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/journal-entries": {
      "get": {
        "tags": [
          "journal-entries-internal"
        ],
        "summary": "Listar lançamentos contábeis",
        "parameters": [
          {
            "name": "company_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "ext_company_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "period",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "2024-01"
            },
            "description": "Filtra pelo mês no formato YYYY-MM"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JournalEntry"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado ou origem não autorizada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-access": "internal"
      },
      "post": {
        "tags": [
          "journal-entries-external"
        ],
        "summary": "Criar lançamento contábil (classificar)",
        "description": "Cria um lançamento em partidas dobradas. Ao informar `flow_transaction_id`,\no lançamento fica vinculado ao lançamento financeiro correspondente,\nmarcando-o como **Classificado**.\n\nDeve sempre ter exatamente 2 linhas (débito e crédito) com o mesmo valor.\n\n\n**Rota externa** — aceita `x-api-key`.\n\n**Validação de `flow_transaction_id`:** Se informado, a transação deve pertencer à mesma empresa (`companyId` ou `extCompanyId`) da requisição. Caso contrário, retorna 403.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "entry_date",
                  "description",
                  "lines"
                ],
                "properties": {
                  "company_id": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Empresa Flow (exclusivo com extCompanyId)"
                  },
                  "ext_company_id": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Empresa externa (exclusivo com companyId)"
                  },
                  "entry_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "description": {
                    "type": "string"
                  },
                  "flow_transaction_id": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true,
                    "description": "Vincula ao lançamento financeiro. **Deve pertencer à mesma empresa da requisição** — caso contrário retorna 403."
                  },
                  "external_ref": {
                    "type": "string",
                    "nullable": true
                  },
                  "lines": {
                    "type": "array",
                    "minItems": 2,
                    "maxItems": 2,
                    "items": {
                      "type": "object",
                      "required": [
                        "account_plan_id",
                        "side",
                        "amount"
                      ],
                      "properties": {
                        "account_plan_id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "side": {
                          "$ref": "#/components/schemas/JournalSide"
                        },
                        "amount": {
                          "type": "number"
                        },
                        "memo": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          }
        },
        "x-access": "external",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ]
      }
    },
...
  }
}